+++ app/src/main/java/uz/axrorxoja/pulbor/App.kt
@@ -1,0 +1,5 @@
+package uz.axrorxoja.pulbor
+
+import android.app.Application
+
+class App : Application()
