+++ app/src/main/java/uz/axrorxoja/pulbor/ui/history/HistoryViewModel.kt
@@ -1,13 +1,13 @@
-package uz.axrorxoja.pulbor.ui.notifications
+package uz.axrorxoja.pulbor.ui.history
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
