+++ app/src/main/java/uz/axrorxoja/pulbor/ui/setting/SettingFragment.kt
@@ -1,4 +1,4 @@
-package uz.axrorxoja.pulbor.ui.notifications
+package uz.axrorxoja.pulbor.ui.setting
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import uz.axrorxoja.pulbor.databinding.FragmentNotificationsBinding
+import uz.axrorxoja.pulbor.databinding.FragmentSettingBinding
 
 class NotificationsFragment : Fragment() {
 
-    private var _binding: FragmentNotificationsBinding? = null
+    private var _binding: FragmentSettingBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -25,7 +25,7 @@ class NotificationsFragment : Fragment() {
         val notificationsViewModel =
                 ViewModelProvider(this).get(NotificationsViewModel::class.java)
 
-        _binding = FragmentNotificationsBinding.inflate(inflater, container, false)
+        _binding = FragmentSettingBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
         val textView: TextView = binding.textNotifications
