+++ app/src/main/java/uz/axrorxoja/pulbor/ui/setting/SettingViewModel.kt
@@ -1,13 +1,13 @@
-package uz.axrorxoja.pulbor.ui.home
+package uz.axrorxoja.pulbor.ui.setting
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class NotificationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
