+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -3,17 +3,17 @@
 
     <item
         android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
-
-    <item
-        android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/ic_dashboard_black_24dp"
+        android:icon="@drawable/ic_dashboard"
         android:title="@string/title_dashboard" />
 
     <item
+        android:id="@+id/navigation_dashboard"
+        android:icon="@drawable/ic_history"
+        android:title="@string/title_history" />
+
+    <item
         android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:icon="@drawable/ic_settings"
+        android:title="@string/title_settings" />
 
 </menu>
