+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="uz.axrorxoja.pulbor.ui.home.HomeFragment"
-        android:label="@string/title_home"
-        tools:layout="@layout/fragment_home" />
-
-    <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="uz.axrorxoja.pulbor.ui.dashboard.DashboardFragment"
+        android:name="uz.axrorxoja.pulbor.ui.dashboard.HomeFragment"
         android:label="@string/title_dashboard"
         tools:layout="@layout/fragment_dashboard" />
 
     <fragment
+        android:id="@+id/navigation_dashboard"
+        android:name="uz.axrorxoja.pulbor.ui.history.DashboardFragment"
+        android:label="@string/title_history"
+        tools:layout="@layout/fragment_history" />
+
+    <fragment
         android:id="@+id/navigation_notifications"
-        android:name="uz.axrorxoja.pulbor.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        android:name="uz.axrorxoja.pulbor.ui.setting.NotificationsFragment"
+        android:label="@string/title_settings"
+        tools:layout="@layout/fragment_setting" />
 </navigation>
